Timemanage/requirements.txt
asgiref==3.2.10
Django==3.1.1
wheel==0.36.2
djangorestframework==3.12.1
mysqlclient==2.0.1
pytz==2020.1
sqlparse==0.4.1
uWSGI==2.0.19.1
